refactor : 당첨 로또의 수를 구하고 수익률을 계산하던 기능을 LottoController에서 WinningLotto로 이동
 - WinningLotto 클래스 생성
 - 당첨 번호와 보너스번호 상태를 관리하도록 변경
 - 당첨 로또의 수와 수익률을 구하는 행위를 관리하도록 변경
